requirements.txt
PyPDF2==3.0.1
openai==1.40.0
anthropic==0.25.0
flask==2.3.3
python-dotenv==1.0.0
httpx==0.25.0